+++ db/migrate/20160315191354_create_users_products.rb
@@ -1,7 +1,7 @@
 class CreateUsersProducts < ActiveRecord::Migration
   def change
     create_table :users_products do |t|
-      t.integer :servings, null: false
+      t.float :servings, null: false
       t.integer :user_id, null: false, foreign_key: true
       t.integer :product_id, null: false, foreign_key: true
       t.date :date_eaten, null: false
